Move logger into a separate file
