+++ 02-express-tutorial/app.js
@@ -1,20 +1,9 @@
 const express = require("express");
 const app = express();
+const logger = require("./logger");
 
-const logger = (req, res, next) => {
-  const method = req.method;
-  const url = req.url;
-  const today = new Date();
-  const date =
-    today.getFullYear() + "-" + (today.getMonth() + 1) + "-" + today.getDate();
-  const time =
-    today.getHours() + ":" + today.getMinutes() + ":" + today.getSeconds();
-  const dateTime = date + " " + time;
-
-  console.log(method, url, dateTime);
-  next();
-};
-app.use(["/", "/about"], logger);
+//app.use(["/", "/about"], logger);
+app.use(logger); //apply to all urls
 
 app.get("/", (req, res) => {
   res.send("Home");
